# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "YTB Backend",
        "description": "Backend API for the YTB Project",
        "contact": {
            "email": "cliff@interfolio.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/v1/sprints": {
            "get": {
                "tags": [
                    "Sprints"
                ],
                "summary": "List all sprints",
                "operationId": "getAllSprints",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Sprint"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            },
            "post": {
                "tags": [
                    "Sprints"
                ],
                "summary": "Create a sprint",
                "operationId": "createSprint",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "start_date",
                                    "end_date"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "The title of the sprint",
                                        "type": "string"
                                    },
                                    "start_date": {
                                        "description": "The sprint start date",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "end_date": {
                                        "description": "The sprint end date",
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successfully created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Sprint"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/v1/sprints/{id}": {
            "get": {
                "tags": [
                    "Sprints"
                ],
                "summary": "Get a sprint",
                "operationId": "getOneSprint",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the sprint to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Sprint"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Sprints"
                ],
                "summary": "Delete a sprint",
                "operationId": "deleteSprint",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the sprint to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully deleted"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/v1/sprints/{$id}": {
            "put": {
                "tags": [
                    "Sprints"
                ],
                "summary": "Update a sprint",
                "operationId": "updateSprint",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the sprint to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "The title of the sprint",
                                        "type": "string"
                                    },
                                    "start_date": {
                                        "description": "The sprint start date",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "end_date": {
                                        "description": "The sprint end date",
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successfully created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Sprint"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "timestamps": {
                "properties": {
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "ErrorItem": {
                "properties": {
                    "field": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "type": "object",
                "example": {
                    "title": "The title field is required."
                }
            },
            "ValidationError": {
                "properties": {
                    "message": {
                        "type": "string"
                    },
                    "errors": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ErrorItem"
                        }
                    }
                },
                "type": "object",
                "example": {
                    "message": "The given data was invalid",
                    "errors": {
                        "title": "The title field is required."
                    }
                }
            },
            "Sprint": {
                "allOf": [
                    {
                        "properties": {
                            "id": {
                                "type": "integer"
                            },
                            "title": {
                                "type": "string"
                            },
                            "start_date": {
                                "type": "string",
                                "format": "date"
                            },
                            "end_date": {
                                "type": "string",
                                "format": "date"
                            }
                        },
                        "type": "object"
                    },
                    {
                        "$ref": "#/components/schemas/timestamps"
                    }
                ]
            }
        }
    },
    "security": [
        []
    ]
}